♻️refactor : 경로 오류, 코드 빠진 것들 추가
