+++ src/pages/Speak.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Component } from "react";
-import Button from './components/Button';
+import Button from '.././components/Common/Button';
 import styled from 'styled-components';
 import Speaker from '../components/Speak/Speaker';
 import {useState , useNavigate} from 'react-router-dom';
 import {useParams} from 'react-router-dom';
-import Questions from './utils/Questions';
-import  ColorCode  from "../utils/palette";
-import Navbar from '../components/Common/Navbar';
+import Questions from '../utils/Questions';
+import  ColorCode  from ".././utils/Palette";
+
 
 const SpeakContainer = styled.div`
    background-color: #F5F5F5;
@@ -30,13 +30,6 @@ const ButtonContainer = styled.div`
       display: flex;
 `;
 
-const Button = styled.button`
-    position : absolute;
-    color : white;
-    
-
-     
-`;
 
 
 
@@ -45,14 +38,14 @@ const  Speak =() =>{
     return (
 
         <SpeakContainer>
-            <Navbar />
+            
             <QuestionContainer>
                <p style={{color : ColorCode.SelectBlue }}>Q1.</p> Tell me about your favorite food.
             </QuestionContainer>
             <Speaker />
             <ButtonContainer>
-                <Button text={"Start"} style = {{color : "white", background:ColorCode.SelectBlue}}/>
-                <Button text={"Done"} />
+                <Button  text={"Start"} style = {{color : "white", background:ColorCode.SelectBlue}}/>
+                <Button text={"Done"} style={{color : "white", background : ColorCode.SelectBlue}} />
             </ButtonContainer>  
         </SpeakContainer>
