+++ src/pages/subject.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { Component } from "react";
 import styled from 'styled-components';
-import Button from './components/Common/Button';
-import Navbar from './components/Common/Navbar';
+import Button from '.././components/Common/Button';
 import {useNavigate} from 'react-router-dom';
+import { useState } from "react";
 import ColorCode from '../utils/Palette';
 
 
@@ -67,7 +67,7 @@ const ChooseTitle = styled.div`
 `;
 
 
-const Button = styled.button`
+const SubButton = styled.button`
   margin: 5px;
   padding: 10px 20px;
   background-color: ${props => (props.active ? ColorCode.SelectBlue : ColorCode.UnSelectBlue)};
@@ -103,7 +103,7 @@ const Subject=() =>{
 
 return (
     <SubjectContainer>
-    <Navbar />
+    
     <Title><p style={{color : "ColorCode.SelectBlue"}}>토킹</p>메이트 시작하기</Title>
     <Documentcontainer>
         <p style={{fontWeight : "bold", size:"20px"}}>사용 설명서</p>
@@ -121,9 +121,9 @@ return (
     <SubjectBox>
         <ChooseTitle>주제를 선택해주세요.</ChooseTitle>
         <SubjectButton>
-            <Button text={"여행"}  active = {activeButton ===0} onClick ={() => handleClick(0)}/>
-            <Button text = {"음식"} active = {activeButton ===1} onCLick = {() => handleClick(1)}/>
-            <Button text ={"가족"} active = {activeButton ===2} onClick = {() => handleClick(2)}/>
+            <SubButton text={"여행"}  active = {activeButton ===0} onClick ={() => handleClick(0)}/>
+            <SubButton text = {"음식"} active = {activeButton ===1} onCLick = {() => handleClick(1)}/>
+            <SubButton text ={"가족"} active = {activeButton ===2} onClick = {() => handleClick(2)}/>
         </SubjectButton>
 
     </SubjectBox>
